#282 ユーザー詳細ではプログレスバーの表示はしない
